move IntervalType closer to its clients
